requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
redis==5.0.4
pydantic==2.7.2
pydantic-settings==2.2.1
openai==1.30.5
python-dotenv==1.0.1
aiohttp==3.9.5